requirements.txt
flask
elasticsearch==8.13.0
sentence-transformers
python-dotenv
